==> analysis/platform/github_team/github_team/2_clean_team_data.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import urllib2
import seaborn
import ConfigParser
import github as gh

#Load team description: I added the countries in manually, so load this to see team members (including )
team_df = pd.read_csv('data/github_team_with_countries.csv', sep='\t')

# Load team repository descriptions
all_team_repos = pd.read_csv('data/github_team_repos.csv', sep='\t')

# Load general event descriptions
all_events = pd.read_csv('data/event_counts_top100k.csv', sep = '\t')

## Clean up countries and cities

team_df['location'] = team_df.location.str.lower().str.replace(',.*', '')

team_df.to_csv('data/github_team_with_countries.csv', sep='\t')

## Clean up team repos names

all_team_repos['repository_name'] = all_team_repos['repository_url'].str.replace('.*/', '')
all_team_repos.to_csv('data/github_team_repos.csv')
